requirements.txt
# DocuMind AI Requirements

# LangChain and Ollama
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.0.1

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.0.0

# Vector Database
chromadb>=0.4.22

# Utilities
tiktoken>=0.5.0
